opt: optimized delete_complete_nodes to not remove one at a time
Since removal of elements in a vec requires moving all subsequent items,
when we remove individual items repeatedly we end up with a worst-case
time complexity of O(n^2). To avoid this we keep track of the start and
end of the range to be deleted, and delete it in one go.