feat: allow concurrent immutable PTableSlices and mutable PieceTable
The main problem was in the function signature itself, by putting the
lifetime parameter in &'ptable self we were telling the compiler that
PTableSlice was borrowing from PieceTable, which is not what we want. We
want the lifetimes to be the same, but for no borrowing to occur. By
removing the self lifetime the compiler was able to figure this out so
now we can have immutable PTableSlices which point to an immutable
PieceTable. This is still unsound for now since the `added` String can
be reallocated on pushes and thus invalidate any pointers in
PTableSlice. This problem will be fixed in future commits.