refactor: remove PTSEnum
the rationale for this is quite complicated, but basically since the
PieceTable holds the added text in a String, the underlying memory
addresses for the text data may change after any changes to that string
due to reallocation. This would invalidate the direct &str reference in
the Contiguous variant of PTSEnum. To solve that I could use a stable
vector of u8 (a vec that doesn't change the memory address of already
existing elements) and it would work, but it would be too much trouble
(for now). So I just decided to remove it and make all slices be the
same, and store a pointer to the String, not to its data, so even if it
reallocates data the pointer will still be valid. This does means that
we have to get through 2 levels of indirection: pointer to String ->
pointer to data which is not ideal but is working for now. I can work on
optimizing after I've laid out all the groundwork, tests and benchmarks