feat: change `Nodes` layout to Struct of Arrays
This is a change that will allow for SIMD operations on piece tables for
faster operations. In order to not change any of the actual logic for
now, There are some APIs in `Nodes` which I don't really want to keep,
but will have to stay for now. Specifically, since `Nodes` is a struct
of array, we want to be iterating only over individual arrays most of
the time, instead of iterating over `Node`s like we're doing right now.

This is quite a big change but all tests are passing so I'm reasonably
confident it's correct.

Also note that insertions at the middle and end have gotten a pretty
consistent 100% performance improvement and start insertions have gotten
a 500%~1000% performance regression (except for large texts, where
there's also a 100% perf improvement). I'm not really sure about the
reason, but I'm not too worried since this SoA design should easily be
more performant with SIMD ops.

Slice operations have gotten somewhat slower as a whole (50%~120%).
Removal operations have gotten consistently faster (100%)