feat: Added simd to find_node
Added SIMD to find_node. In order to optimize for max performance, Nodes
are first searched in batches of 64, then 8, then 1. The naming schemes
are a bit awkward I think, but they're okay for now. Getting all the
tests to pass was surprisinly easy, which is a good sign, I think. It
means that my interfaces are well and dandy and changing the
implementation for underneath the code doesn't break it (if it still
gives the same results).